+++ src/test/java/bowling/domain/frame/FramesTest.java
@@ -1,6 +1,6 @@
 package bowling.domain.frame;
 
-import bowling.domain.exception.InvalidFramesException;
+import bowling.exception.InvalidFramesException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
