+++ src/test/java/bowling/domain/state/FirstBowlTest.java
@@ -1,7 +1,7 @@
 package bowling.domain.state;
 
 import bowling.domain.Pins;
-import bowling.domain.exception.InvalidPitchException;
+import bowling.exception.InvalidPitchException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
